extract tests to their own class - add one more test with multiple sources
