+++ Tests/SonarScanner.MSBuild.PreProcessor.Test/PreProcessorTests.cs
@@ -23,7 +23,7 @@ using SonarScanner.MSBuild.PreProcessor.Roslyn.Model;
 namespace SonarScanner.MSBuild.PreProcessor.Test;
 
 [TestClass]
-public class PreProcessorTests
+public partial class PreProcessorTests
 {
     public TestContext TestContext { get; set; }
 
@@ -418,100 +418,6 @@ public class PreProcessorTests
         AssertExpectedServerSetting(actualConfig, "server.key", "server value 1");
         AssertExpectedServerSetting(actualConfig, "shared.key1", "server shared value 1");
         AssertExpectedServerSetting(actualConfig, "shared.CASING", "server upper case value");
-    }
-
-    [TestMethod]
-    public async Task Execute_WritesTelemetry_SetViaCLI()
-    {
-        var args = new List<string>(CreateArgs())
-        {
-            "/d:sonar.scanner.scanAll=false"
-        };
-        using var scope = new TestScope(TestContext);
-        var factory = new MockObjectFactory();
-        var settings = factory.ReadSettings();
-        var preProcessor = new PreProcessor(factory, new ConsoleLogger(false));
-
-        var success = await preProcessor.Execute(args);
-
-        success.Should().BeTrue("Expecting the pre-processing to complete successfully");
-        Directory.GetFiles(settings.SonarOutputDirectory).Select(Path.GetFileName).Should().Contain(FileConstants.TelemetryFileName);
-        File.ReadAllText(Path.Combine(settings.SonarOutputDirectory, FileConstants.TelemetryFileName))
-            .Should()
-            .BeEquivalentTo("""
-            {"s4net.params.sonar_scanner_scanAll.value":"false"}
-            {"s4net.params.sonar_scanner_scanAll.source":"CLI"}
-
-            """);
-    }
-
-    [TestMethod]
-    public async Task Execute_WritesTelemetry_SetViaServer()
-    {
-        using var scope = new TestScope(TestContext);
-        var factory = new MockObjectFactory(serverProperties: new Dictionary<string, string> { { "sonar.scanner.scanAll", "false" } });
-        var settings = factory.ReadSettings();
-        var preProcessor = new PreProcessor(factory, new ConsoleLogger(false));
-
-        var success = await preProcessor.Execute(CreateArgs());
-
-        success.Should().BeTrue("Expecting the pre-processing to complete successfully");
-        Directory.GetFiles(settings.SonarOutputDirectory).Select(Path.GetFileName).Should().Contain(FileConstants.TelemetryFileName);
-        File.ReadAllText(Path.Combine(settings.SonarOutputDirectory, FileConstants.TelemetryFileName))
-            .Should()
-            .BeEquivalentTo("""
-            {"s4net.params.sonar_scanner_scanAll.value":"false"}
-            {"s4net.params.sonar_scanner_scanAll.source":"SQ_GLOBAL_SETTINGS"}
-
-            """);
-    }
-
-    [TestMethod]
-    public async Task Execute_WritesTelemetry_SetViaAnalysisXml()
-    {
-        var analysisXmlPath = CreateAnalysisXml(TestContext.TestRunDirectory, new Dictionary<string, string> { { "sonar.scanner.scanAll", "false" } });
-        var args = new List<string>(CreateArgs())
-        {
-            $"/s:{analysisXmlPath}"
-        };
-        using var scope = new TestScope(TestContext);
-        var factory = new MockObjectFactory();
-        var settings = factory.ReadSettings();
-        var preProcessor = new PreProcessor(factory, new ConsoleLogger(false));
-
-        var success = await preProcessor.Execute(args);
-
-        success.Should().BeTrue("Expecting the pre-processing to complete successfully");
-        Directory.GetFiles(settings.SonarOutputDirectory).Select(Path.GetFileName).Should().Contain(FileConstants.TelemetryFileName);
-        File.ReadAllText(Path.Combine(settings.SonarOutputDirectory, FileConstants.TelemetryFileName))
-            .Should()
-            .BeEquivalentTo("""
-            {"s4net.params.sonar_scanner_scanAll.value":"false"}
-            {"s4net.params.sonar_scanner_scanAll.source":"SONARQUBE_ANALYSIS_XML"}
-
-            """);
-    }
-
-    [TestMethod]
-    public async Task Execute_WritesTelemetry_SetViaEnvVariable()
-    {
-        using var scope = new TestScope(TestContext);
-        var factory = new MockObjectFactory();
-        var settings = factory.ReadSettings();
-        Environment.SetEnvironmentVariable("SONARQUBE_SCANNER_PARAMS", """{"sonar.scanner.scanAll": "false"}""");
-        var preProcessor = new PreProcessor(factory, new ConsoleLogger(false));
-
-        var success = await preProcessor.Execute(CreateArgs());
-
-        success.Should().BeTrue("Expecting the pre-processing to complete successfully");
-        Directory.GetFiles(settings.SonarOutputDirectory).Select(Path.GetFileName).Should().Contain(FileConstants.TelemetryFileName);
-        File.ReadAllText(Path.Combine(settings.SonarOutputDirectory, FileConstants.TelemetryFileName))
-            .Should()
-            .BeEquivalentTo("""
-            {"s4net.params.sonar_scanner_scanAll.value":"false"}
-            {"s4net.params.sonar_scanner_scanAll.source":"SONARQUBE_SCANNER_PARAMS"}
-
-            """);
     }
 
     private static IEnumerable<string> CreateArgs(string organization = null, Dictionary<string, string> properties = null)
